Game

Documentation

data Player

It represents player.

Constructors

PlayerB

PlayerW

data Cell

It represents the state of the cell.

Constructors

Empty

Full Player

Text1

Text2

Text3

Text4

Text5

Text6

Text7

Text8

data State

It represents the state of the game.

Constructors

Running

GameOver (Maybe Player)

type Board = Array (Int, Int) Cell

list of list of cells.

data Game

this data set represents the state of game at any point of time.

Constructors

Game

Fields

gameBoard :: Board

gamePlayer :: Player

gameState :: State

n :: Int

number of rows or columns

screenWidth :: Int

screen width

screenWidth1 :: Int

screen width1

screenHeight :: Int

screen height

screenHeight1 :: Int

screen height1

cellWidth :: Float

This cellWidth defines the width of the cell

cellHeight :: Float

This cellHeight defines the height of the cell

initialGame :: Game

This initialGame states the initial conditions of the game |first player is PlayerB and game state is running.