Logic

Documentation

m :: Int

isCoordCorrect :: (Int, Int) -> Bool

This isCoordCorrect function checks if the cell coord in the board range or not.

isCoordAdjacent

Arguments

:: Game
takes present game state as input parameter.

-> (Int, Int)
position of the newly placed cell as input

-> Bool
returns true or false.

This isCoordAdjacent function checks if the adjacent position contains discs of the opposite player.

isendSymbolDown

Arguments

:: Game
takes state of game as input parameter

-> (Int, Int)
position of the colour cell as input

-> Bool
returns true or false

This isendSymbolDown function checks if on the down side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isDown

Arguments

:: Game
this takes input as game state

-> (Int, Int)
and position of the new cell

-> Bool
returns a true or false.

This isDown symbol returns true if we are able to flank the opponents discs in the down direction else false

isendSymbolUp :: Game -> (Int, Int) -> Bool

This isendSymbolUp function checks if on the up side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isUp :: Game -> (Int, Int) -> Bool

This isUp symbol returns true if we are able to flank the opponents discs in the up direction else false

isendSymbolLeft :: Game -> (Int, Int) -> Bool

This isendSymbolLeft function checks if on the left side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isLeft :: Game -> (Int, Int) -> Bool

This isLeft symbol returns true if we are able to flank the opponents discs in the left direction else false

isendSymbolRight :: Game -> (Int, Int) -> Bool

This isendSymbolRight function checks if on the right side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isRight :: Game -> (Int, Int) -> Bool

This isRight symbol returns true if we are able to flank the opponents discs in the right direction else false

isendSymbolNW :: Game -> (Int, Int) -> Bool

This isendSymbolNW function checks if on the north west side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isNW :: Game -> (Int, Int) -> Bool

This isNW symbol returns true if we are able to flank the opponents discs in the north west direction else false

isendSymbolNE :: Game -> (Int, Int) -> Bool

This isendSymbolNE function checks if on the north east side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isNE :: Game -> (Int, Int) -> Bool

This isNE symbol returns true if we are able to flank the opponents discs in the north east direction else false

isendSymbolSE :: Game -> (Int, Int) -> Bool

This isendSymbolSE function checks if on the south east side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isSE :: Game -> (Int, Int) -> Bool

This isSE symbol returns true if we are able to flank the opponents discs in the south east direction else false

isendSymbolSW :: Game -> (Int, Int) -> Bool

This isendSymbolSW function checks if on the south west side we have a disc of same color |to form a straight line enclosing discs of opposite color to flank them.

isSW :: Game -> (Int, Int) -> Bool

This isSW symbol returns true if we are able to flank the opponents discs in the south west direction else false

genUp :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenUp function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the up direction

upIndex :: Game -> (Int, Int) -> (Int, Int)

This upIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the up direction

flipUp :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipUp function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the up direction.

checkflipUp :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipUp function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genDown :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenDown function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the down direction

downIndex :: Game -> (Int, Int) -> (Int, Int)

This downIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the down direction

flipDown :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipDown function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the down direction.

checkflipDown :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipDown function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genLeft :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenLeft function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the left direction

leftIndex :: Game -> (Int, Int) -> (Int, Int)

This leftIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the left direction

flipLeft :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipLeft function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the left direction.

checkflipLeft :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipLeft function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genRight :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenRight function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the right direction

rightIndex :: Game -> (Int, Int) -> (Int, Int)

This rightIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the right direction

flipRight :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipRight function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the right direction.

checkflipRight :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipRight function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genNW :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenNW function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the north west direction

nwIndex :: Game -> (Int, Int) -> (Int, Int)

This nwIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the north west direction

flipNW :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipNW function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the north west direction.

checkflipNW :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipNW function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genNE :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenNE function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the north east direction

neIndex :: Game -> (Int, Int) -> (Int, Int)

This neIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the north east direction

flipNE :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipNE function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the north east direction.

checkflipNE :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipNE function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genSW :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenSW function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the south west direction

swIndex :: Game -> (Int, Int) -> (Int, Int)

This swIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the south west direction

flipSW :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipSW function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the south west direction.

checkflipSW :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipSW function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

genSE :: ((Int, Int), (Int, Int)) -> [(Int, Int)]

This isgenSE function returns a list of [(Int,Int)] which is a list of cellcoords to be flipped in the south east direction

seIndex :: Game -> (Int, Int) -> (Int, Int)

This seIndex function returns (Int,Int) which is an extreme end of cellcoords to be flipped in the south east direction

flipSE :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This flipSE function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped along with the player type for each cellcoord in the south east direction.

checkflipSE :: Game -> (Int, Int) -> [((Int, Int), Cell)]

This checkflipSE function returns a list of [((Int,Int),Cell)] which is a list of cellcoords to be flipped in the up direction or [] if no flipping is possible in this direction.

switchPlayer

Arguments

:: Game
takes sate of game as input

-> Game
ouputs the state of the game.

This switchPlayer function returns a game by taking input a game and switching the game Player

calculate

Arguments

:: Game
this takes the present game state as input parameter.

-> Cell
this takes present state of the cell

-> [(Int, Int)]
list of positions

-> Int
returns the Int which is frequency.

This calculate function takes a game, cell and a list and returns the frequency of the cell in the game Board

checkGameOver

Arguments

:: Game
takes present game state as input parameter.

-> Game
returns the new game state

This checkGameOver function take a game and checks if the game is over and returns a new game with gamestate equal to gameOver

checkValidMoves :: Game -> [(Int, Int)] -> Bool

This checkValidMoves function takes a Game and a cell coord and checks if it is a valid move to place a disc at that cell coord.

isValidMovesLeft :: Game -> Bool

This isValidMovesLeft function takes a Game and returns a bool value if any possible move is available or not.

checkValidPositon

Arguments

:: Game
the Gameinput parametr

-> (Int, Int)
the position as input parameter.

-> Bool
returns true or false.

This checkValidPosition checks whether it is possible to place on cell coord (Int,Int)

setPoints :: Game -> (Int, Int) -> Int

Thiss setPoints returns a int which is no of flips of opposite player which would occur on placing on the cellCoord input

findIndex :: Int -> [Int] -> Int -> (Int, Int)

This findIndex function returns a cellCoords which corresponds to the index of last input in the input list

findMaximum :: [Int] -> (Int, Int)

This findMaximum function returns a cell coord corresponding to the maximum value int the input list

getPos :: Game -> (Int, Int)

This getPos function returns the cellcoord for the motion of bot.

botMove :: Game -> (Int, Int)

This function return the position of bot motion.

botMotion :: Game -> Game

This botMotion function takes a game applies the bot motion to it and returns the next state of game.

playerTurn :: Game -> (Int, Int) -> Game

This function checks for valid input by user, places its disc, flipps opponent discs and switches player | This function also calls for botMotion if the player is PlayerB.

mousePosAsCellCoord :: (Float, Float) -> (Int, Int)

This mousePosAsCellCoord function takes the click position of user, maps it to a cell index and retuns the coordinates.

transformGame :: Event -> Game -> Game

This function transforms the state of the game according to the events done by the user.