Rendering

Description

this is responsible for the graphics.

Documentation

boardGridColor :: Color

assigning different colours to different things on the board.

playerBColor :: Color

playerWColor :: Color

stringColor :: Color

playerWboardColor :: Color

tieColor :: Color

boardAsRunningPicture :: Game -> Board -> Picture

outcomeColor :: Maybe Player -> Color

snapPictureToCell :: (Integral t, Integral t1) => Picture -> (t, t1) -> Picture

takes picture and coordinates and translates accordingly.

bCell :: Picture

for creating a black cell.

wCell :: Picture

for creating a white cell.

sCell1 :: Picture

This sCell1 display text from 1..8.

sCell2 :: Picture

This sCell2 display text from 1..8.

sCell3 :: Picture

This sCell3 display text from 1..8.

sCell4 :: Picture

This sCell4 display text from 1..8.

sCell5 :: Picture

This sCell5 display text from 1..8.

sCell6 :: Picture

This sCell6 display text from 1..8.

sCell7 :: Picture

This sCell7 display text from 1..8.

sCell8 :: Picture

This sCell8 display text from 1..8.

cellsOfBoard :: Board -> Cell -> Picture -> Picture

This function cellsOfBoard renders differnt pictures.

bCellsOfBoard :: Board -> Picture

wCellsOfBoard :: Board -> Picture

s1CellsOfBoard :: Board -> Picture

s2CellsOfBoard :: Board -> Picture

s3CellsOfBoard :: Board -> Picture

s4CellsOfBoard :: Board -> Picture

s5CellsOfBoard :: Board -> Picture

s6CellsOfBoard :: Board -> Picture

s7CellsOfBoard :: Board -> Picture

s8CellsOfBoard :: Board -> Picture

boardGrid :: Picture

This boardGrid for drawing grid to our board.

boardAsPicture :: Board -> Picture

this returns the board combiation with white cells black cells and grid .

boardAsGameOverPicture :: Maybe Player -> Board -> Picture

gameAsPicture :: Game -> Picture

This function gameAsPicture takes presents game state as input parameter and renders it.